# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fathom",
    "version": "1.0.0",
    "description": "Macroeconomic data API powered by FRED (Federal Reserve Economic Data). 816,000+ time series covering GDP, inflation, employment, interest rates, money supply, and more.",
    "x-guidance": "Use /api/popular to discover available indicators. Use /api/search to find specific series by keyword. Use /api/observations to fetch data values. Use /api/snapshot for a dashboard overview. Use /api/compare to compare multiple series. Series IDs are uppercase codes like GDP, UNRATE, CPIAUCSL, FEDFUNDS, DGS10.",
    "contact": {
      "email": "admin@orbonomy.xyz"
    },
    "x-logo": {
      "url": "https://fathom.vercel.app/favicon.png"
    }
  },
  "servers": [
    {
      "url": "https://fathom.vercel.app"
    }
  ],
  "paths": {
...
    "/api/search": {
      "post": {
        "operationId": "search",
        "summary": "Search economic time series",
        "description": "Search across 816,000+ FRED economic time series by keyword.",
        "tags": [
          "Discovery"
        ],
        "security": [],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.020000"
          },
          "protocols": [
            {
              "x402": {
                "version": 2,
                "asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                "network": "eip155:8453"
              }
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Search keywords (e.g. 'inflation rate', 'housing starts')"
                  },
                  "limit": {
                    "type": "string",
                    "description": "Max results (default: 10)"
                  }
                },
                "required": [
                  "query"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search economic time series",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string"
                    },
                    "query": {
                      "type": "string"
                    },
                    "count": {
                      "type": "number"
                    },
                    "series": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  },
                  "required": [
                    "source"
                  ]
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
...
    "/api/releases": {
      "post": {
        "operationId": "releases",
        "summary": "Data release schedule",
        "description": "Upcoming Federal Reserve economic data releases.",
        "tags": [
          "Metadata"
        ],
        "security": [],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.020000"
          },
          "protocols": [
            {
              "x402": {
                "version": 2,
                "asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                "network": "eip155:8453"
              }
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "limit": {
                    "type": "string",
                    "description": "Max results (default: 20)"
                  },
                  "dates": {
                    "type": "boolean",
                    "description": "If true, return dates only"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Data release schedule",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string"
                    },
                    "releases": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  },
                  "required": [
                    "source"
                  ]
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
...
    "/api/tags": {
      "post": {
        "operationId": "tags",
        "summary": "Browse tags",
        "description": "Browse all FRED tags. Tags describe data attributes like 'monthly', 'gdp', 'inflation', 'nsa'.",
        "tags": [
          "Discovery"
        ],
        "security": [],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.010000"
          },
          "protocols": [
            {
              "x402": {
                "version": 2,
                "asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                "network": "eip155:8453"
              }
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "limit": {
                    "type": "string",
                    "description": "Max results (default: 50)"
                  },
                  "search": {
                    "type": "string",
                    "description": "Filter tags by keyword"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Browse tags",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string"
                    },
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  },
                  "required": [
                    "source"
                  ]
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
    "/api/tag-series": {
      "post": {
        "operationId": "tagseries",
        "summary": "Series by tag",
        "description": "Find economic data series matching a specific tag (e.g. 'gdp', 'monthly', 'inflation').",
        "tags": [
          "Discovery"
        ],
        "security": [],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.020000"
          },
          "protocols": [
            {
              "x402": {
                "version": 2,
                "asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                "network": "eip155:8453"
              }
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tag": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Tag name (e.g. gdp, monthly, inflation, nsa)"
                  },
                  "limit": {
                    "type": "string",
                    "description": "Max results (default: 20)"
                  }
                },
                "required": [
                  "tag"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Series by tag",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string"
                    },
                    "tag": {
                      "type": "string"
                    },
                    "series": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  },
                  "required": [
                    "source"
                  ]
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
...
    "/api/sources": {
      "post": {
        "operationId": "sources",
        "summary": "Data sources",
        "description": "List all FRED data sources (Federal Reserve, BLS, Census, BEA, etc.).",
        "tags": [
          "Metadata"
        ],
        "security": [],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.010000"
          },
          "protocols": [
            {
              "x402": {
                "version": 2,
                "asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                "network": "eip155:8453"
              }
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Data sources",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string"
                    },
                    "sources": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  },
                  "required": [
                    "source"
                  ]
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
...
    "/api/series-tags": {
      "post": {
        "operationId": "seriestags",
        "summary": "Tags for a series",
        "description": "Get all tags associated with a specific FRED series.",
        "tags": [
          "Metadata"
        ],
        "security": [],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.010000"
          },
          "protocols": [
            {
              "x402": {
                "version": 2,
                "asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                "network": "eip155:8453"
              }
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "seriesId": {
                    "type": "string",
                    "minLength": 1,
                    "description": "FRED series ID"
                  }
                },
                "required": [
                  "seriesId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tags for a series",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string"
                    },
                    "series_id": {
                      "type": "string"
                    },
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  },
                  "required": [
                    "source"
                  ]
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
    "/api/search-tags": {
      "post": {
        "operationId": "searchtags",
        "summary": "Search tags",
        "description": "Search FRED tags by keyword.",
        "tags": [
          "Discovery"
        ],
        "security": [],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.010000"
          },
          "protocols": [
            {
              "x402": {
                "version": 2,
                "asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                "network": "eip155:8453"
              }
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Search keyword"
                  },
                  "limit": {
                    "type": "string",
                    "description": "Max results (default: 20)"
                  }
                },
                "required": [
                  "query"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search tags",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string"
                    },
                    "query": {
                      "type": "string"
                    },
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  },
                  "required": [
                    "source"
                  ]
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
...
  }
}